# Add optional profiling context managers and annotate serialization stages