--- a/datason/__init__.py
+++ b/datason/__init__.py
@@ -41,7 +41,7 @@ from .api import (
     get_api_info,
     help_api,
     load,  # Enhanced default (smart parsing, datetime support)
-    load_basic,
+    # NOTE: load_basic redefined below for benchmark integration
     # NEW: File I/O operations integrated with modern API
     load_basic_file,
     load_json,  # JSON compatibility (exact stdlib behavior)
@@ -193,18 +193,74 @@ from .integrity import (  # noqa: F401
 )
 from .validation import serialize_marshmallow, serialize_pydantic  # noqa: F401
 
+# =============================================================================
+# PROFILING AND BENCHMARKING SUPPORT
+# =============================================================================
+
+# Profile sink for external benchmarking tools (e.g., datason-benchmark)
+profile_sink = None  # Will be set by external tools as a list to append events
+
 
 def set_profile_sink(sink: Optional[Callable[[dict], None]]) -> None:
     """Register a callback to receive profiling timings.
 
     The callable will be invoked with a dictionary of stage timings whenever
-    profiling is enabled and a top-level operation completes.  Importing is
+    profiling is enabled and a top-level operation completes. Importing is
     deferred to avoid any overhead when profiling is disabled.
-    """
 
+    Args:
+        sink: Callable that receives timing dictionaries, or None to disable
+    """
     from ._profiling import set_profile_sink as _set_profile_sink
 
     _set_profile_sink(sink)
+
+
+# Rust availability detection (will be implemented in codex/add-rust-core-mvp-for-datason)
+RUST_AVAILABLE = False
+
+
+# Core APIs for benchmarking and Rust exploration
+def save_string(obj: Any) -> str:
+    """Serialize object to JSON string.
+
+    Primary API expected by datason-benchmark for measuring serialization performance.
+
+    Args:
+        obj: Object to serialize
+
+    Returns:
+        JSON string representation
+
+    Example:
+        >>> result = datason.save_string({"key": "value"})
+        >>> assert isinstance(result, str)
+    """
+    # Use dumps_json for direct string output
+    from .api import dumps_json
+
+    return dumps_json(obj)
+
+
+def load_basic(json_str: str) -> Any:
+    """Parse JSON string to Python object.
+
+    Primary API expected by datason-benchmark for measuring deserialization performance.
+
+    Args:
+        json_str: JSON string to parse
+
+    Returns:
+        Parsed Python object
+
+    Example:
+        >>> obj = datason.load_basic('{"key": "value"}')
+        >>> assert obj == {"key": "value"}
+    """
+    # Use loads_json for direct string parsing
+    from .api import loads_json
+
+    return loads_json(json_str)
 
 
 def _get_version() -> str:
@@ -244,7 +300,12 @@ __all__ = [  # noqa: RUF022
     "load",  # Enhanced file reading with smart parsing
     "loads",  # Enhanced string parsing with smart features
     "serialize",  # Enhanced serialization (returns dict)
+    # Profiling and benchmarking support
     "set_profile_sink",  # Debug profiling sink registration
+    "profile_sink",  # External profiling event list
+    "RUST_AVAILABLE",  # Rust backend availability flag
+    "save_string",  # Core benchmarking API (serialize to string)
+    "load_basic",  # Core benchmarking API (parse from string)
     # JSON Compatibility API (for stdlib replacement)
     "dump_json",  # Exact json.dump() behavior
     "dumps_json",  # Exact json.dumps() behavior (returns string)