--- a/datason/_profiling.py
+++ b/datason/_profiling.py
@@ -1,96 +1,117 @@
-# Lightweight profiling utilities for optional stage timing.
 """Internal profiling utilities for optional stage timing.
 
 These helpers provide minimal-overhead timing that can be enabled via the
-``DATASON_PROFILE`` environment variable.  When the variable is set to ``"1"``,
-code wrapped in :func:`profile_run` and :func:`stage` will record wall clock
-execution times for named stages.  When the variable is unset or ``"0"`` these
-functions become near no-ops and the overhead is intended to be effectively
-zero.
-
-The collected timings can be forwarded to an optional *profile sink* – a
-callable that receives the dictionary of stage timings when the top level run
-completes.  The sink can be configured via :func:`set_profile_sink`.
+``DATASON_PROFILE`` environment variable. When enabled, blocks wrapped in
+``profile_run`` and ``stage`` accumulate wall-clock timings for named stages.
+When disabled they devolve to no-ops with negligible overhead. Collected
+timings can be forwarded to an optional *profile sink* – a callable that
+receives the stage dictionary when the top level run completes.
 """
 
 from __future__ import annotations
 
+import contextvars
 import os
 import time
-from contextlib import contextmanager
-from typing import Callable, Dict, Optional
-import contextvars
+from typing import Callable
 
 # Evaluate the flag once for minimal overhead in hot paths
 ENABLED = os.getenv("DATASON_PROFILE") == "1"
 
 # Context variable storing the timing dictionary for the current run
-_current_timings: contextvars.ContextVar[Optional[Dict[str, float]]] = contextvars.ContextVar(
-    "datason_profile_timings", default=None
+_current_timings: contextvars.ContextVar[dict[str, float] | None] = (
+    contextvars.ContextVar("datason_profile_timings", default=None)
 )
 
 # Optional sink callable invoked with timings at the end of a profiling run
-_profile_sink: Optional[Callable[[Dict[str, float]], None]] = None
+_profile_sink: Callable[[dict[str, float]], None] | None = None
 
 
-def set_profile_sink(sink: Optional[Callable[[Dict[str, float]], None]]) -> None:
+def set_profile_sink(sink: Callable[[dict[str, float]], None] | None) -> None:
     """Register a callable that receives timing dictionaries.
 
-    The *sink* is invoked once for each top-level :func:`profile_run` when
-    profiling is enabled.  Passing ``None`` removes the sink.
+    The *sink* is invoked once for each top-level ``profile_run`` when profiling
+    is enabled. Passing ``None`` removes the sink.
     """
 
     global _profile_sink
     _profile_sink = sink
 
 
-@contextmanager
-def profile_run() -> Dict[str, float]:
-    """Create a profiling run and yield the timings dictionary.
+class _NullProfileRun:
+    def __enter__(self) -> dict[str, float]:  # noqa: D401 - trivial context
+        return {}
 
-    When profiling is disabled this yields an empty dictionary.  When enabled it
-    yields a dictionary that will be populated by :func:`stage` contexts.
-    """
+    def __exit__(self, exc_type, exc, tb) -> bool:  # noqa: D401 - trivial context
+        return False
 
-    if not ENABLED:
-        yield {}
-        return
 
-    timings: Dict[str, float] = {}
-    token = _current_timings.set(timings)
-    try:
-        yield timings
-    finally:
-        _current_timings.reset(token)
-        if _profile_sink is not None:
+class _ProfileRun:
+    def __init__(self) -> None:
+        self.timings: dict[str, float] = {}
+        self.token: contextvars.Token | None = None
+
+    def __enter__(self) -> dict[str, float]:  # noqa: D401 - context manager
+        self.token = _current_timings.set(self.timings)
+        return self.timings
+
+    def __exit__(self, exc_type, exc, tb) -> bool:  # noqa: D401 - context manager
+        if self.token is not None:
+            _current_timings.reset(self.token)
+        sink = _profile_sink
+        if sink is not None:
             try:
-                _profile_sink(timings)
+                sink(self.timings)
             except Exception:
                 # Profiling must never interfere with normal execution
                 pass
+        return False
 
 
-@contextmanager
-def stage(name: str):
-    """Measure a named stage within a :func:`profile_run`.
+class _NullStage:
+    def __enter__(self) -> None:  # noqa: D401 - trivial context
+        return None
 
-    The collected timings are stored in the current run's dictionary.  When
-    profiling is disabled this context manager adds virtually no overhead.
-    """
+    def __exit__(self, exc_type, exc, tb) -> bool:  # noqa: D401 - trivial context
+        return False
+
+
+class _Stage:
+    def __init__(self, name: str) -> None:
+        self.name = name
+        self.timings: dict[str, float] | None = None
+        self.start: float = 0.0
+
+    def __enter__(self) -> None:  # noqa: D401 - context manager
+        self.timings = _current_timings.get()
+        if self.timings is not None:
+            self.start = time.perf_counter()
+        return None
+
+    def __exit__(self, exc_type, exc, tb) -> bool:  # noqa: D401 - context manager
+        timings = self.timings
+        if timings is not None:
+            end = time.perf_counter()
+            timings[self.name] = timings.get(self.name, 0.0) + (end - self.start)
+        return False
+
+
+_NULL_PROFILE_RUN = _NullProfileRun()
+_NULL_STAGE = _NullStage()
+
+
+def profile_run() -> _ProfileRun | _NullProfileRun:
+    """Create a profiling run and yield the timings dictionary."""
 
     if not ENABLED:
-        yield
-        return
+        return _NULL_PROFILE_RUN
+    return _ProfileRun()
 
-    timings = _current_timings.get()
-    if timings is None:
-        # No active profiling run
-        yield
-        return
 
-    start = time.perf_counter()
-    try:
-        yield
-    finally:
-        end = time.perf_counter()
-        timings[name] = timings.get(name, 0.0) + (end - start)
+def stage(name: str) -> _Stage | _NullStage:
+    """Measure a named stage within a ``profile_run``."""
+
+    if not ENABLED:
+        return _NULL_STAGE
+    return _Stage(name)
+