--- a/tests/unit/test_profiling.py
+++ b/tests/unit/test_profiling.py
@@ -8,7 +8,7 @@ ROOT = Path(__file__).resolve().parents[2]
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-import datason
+import datason  # noqa: E402
 
 
 def test_profile_run_disabled(monkeypatch):