## Summary
- Introduce `_profiling` module with `profile_run` and `stage` context managers gated by `DATASON_PROFILE`
- Expose `set_profile_sink` helper and wrap `serialize`/`load_basic` to emit timing data
- Annotate core `serialize` and `deserialize` with named stages for fine-grained performance insight

## Testing
- `pytest tests/unit/test_profiling.py -q`

------
https://chatgpt.com/codex/tasks/task_e_689f52713748832589f1e2c4fb638e85