@danielendler 2025-08-15
# 🚀 DataSON PR Performance Analysis

**PR #79** | **Commit**: `088c539575c225ac9c8ca5a879bbc3c05f2f793c`

## 📊 Benchmark Results

**Suite**: pr_optimized | **Tests Run**: 5 | **Success Rate**: 100.0%

### 🎯 DataSON Performance Summary

| Metric | Result | Benchmark Details | Status |
|--------|--------|-------------------|--------|
| Serialization (avg) | **1.127 ms** | 5 measurements across 5 scenarios | ✅ |
| Deserialization (avg) | **0.000 ms** | 0 measurements | ❌ |
| Success Rate | **100.0%** | 5 scenarios tested with 5 iterations each | ✅ |
| Performance Range | 0.120 - 3.676 ms | Min to max serialization times | ✅ |

### 📋 Test Scenarios

| Scenario | Status | Serialization | Deserialization |
|----------|--------|---------------|-----------------|
| Web Api Response | ✅ Passed | 1.405 ms | N/A |
| Ml Training Batch | ✅ Passed | 3.676 ms | N/A |
| Financial Transaction | ✅ Passed | 0.179 ms | N/A |
| Mixed Types Challenge | ✅ Passed | 0.253 ms | N/A |
| Security Pii Test | ✅ Passed | 0.12 ms | N/A |

## ℹ️ Baseline Status

No baseline available for comparison. Future PRs will compare against this baseline.

## ✅ Status: Ready for Review

All benchmarks passed! No significant performance regressions detected.

---
*Generated by [datason-benchmarks](https://github.com/danielendler/datason-benchmarks) • Comprehensive Performance Analysis*